リファクタリング: ファイルをコンポーネントに分割
