+++ video-overlay/app/components/Faces.js
@@ -1,0 +1,15 @@
+import Face from "../components/Face";
+
+export default function Faces({ faces, shows }) {
+  if ( faces == null ) return null;
+  return (
+    <g>
+      {faces.map((face, index) => (
+        <Face
+            key={index}
+            face={face}
+            shows={shows} />
+      ))}
+    </g>
+  );
+}
