+++ video-overlay/app/pages/index.js
@@ -1,263 +1,9 @@
 import Head from "next/head";
-import React, { useState, useEffect, useRef, useCallback } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 
-function BoundingBox({ x1, y1, x2, y2, color, size = 0.2, strokeOpacity = 0.7, strokeWidth = 3, border = 2 }) {
-  const dx = (x2 - x1) * size;
-  const dy = (y2 - y1) * size;
-  return (
-    <g>
-      <path
-          d={
-            `M${x1 - border},${y1 + dy} L${x1 - border},${y1 - border}  L${x1 + dx},${y1 - border}`
-            + ` M${x2 + border},${y1 + dy} L${x2 + border},${y1 - border}  L${x2 - dx},${y1 - border}`
-            + ` M${x1 - border},${y2 - dy} L${x1 - border},${y2 + border}  L${x1 + dx},${y2 + border}`
-            + ` M${x2 + border},${y2 - dy} L${x2 + border},${y2 + border}  L${x2 - dx},${y2 + border}`
-          }
-          stroke="white"
-          strokeOpacity={strokeOpacity}
-          strokeWidth={border}
-          fill="none" />
-      <path
-          d={
-            `M${x1 + border},${y1 + dy} L${x1 + border},${y1 + border}  L${x1 + dx},${y1 + border}`
-            + ` M${x2 - border},${y1 + dy} L${x2 - border},${y1 + border}  L${x2 - dx},${y1 + border}`
-            + ` M${x1 + border},${y2 - dy} L${x1 + border},${y2 - border}  L${x1 + dx},${y2 - border}`
-            + ` M${x2 - border},${y2 - dy} L${x2 - border},${y2 - border}  L${x2 - dx},${y2 - border}`
-          }
-          stroke="white"
-          strokeOpacity={strokeOpacity}
-          strokeWidth={border}
-          fill="none" />
-      <path
-          d={
-            `M${x1},${y1 + dy} L${x1},${y1}  L${x1 + dx},${y1}`
-            + ` M${x2},${y1 + dy} L${x2},${y1}  L${x2 - dx},${y1}`
-            + ` M${x1},${y2 - dy} L${x1},${y2}  L${x1 + dx},${y2}`
-            + ` M${x2},${y2 - dy} L${x2},${y2}  L${x2 - dx},${y2}`
-          }
-          stroke={color}
-          strokeOpacity={strokeOpacity}
-          strokeWidth={strokeWidth}
-          fill="none" />
-    </g>
-  );
-}
-
-function ScoreBar({ score, x1, y1, x2, y2, color, height = 4, padding = 5, strokeWidth = 3, strokeOpacity = 0.7, fillOpacity = 0.8 }) {
-  const width = x2 - x1;
-  return (
-    <g>
-      <rect
-          x={x1 + padding}
-          y={y1 + padding}
-          width={width - padding * 2}
-          height={height}
-          stroke="white"
-          strokeWidth={strokeWidth}
-          strokeOpacity={strokeOpacity}
-          fill="white"
-          fillOpacity={fillOpacity} />
-      <rect
-          x={x1 + padding}
-          y={y1 + padding}
-          width={(width - padding * 2) * score}
-          height={height}
-          fill={color}
-          fillOpacity={fillOpacity} />
-    </g>
-  );
-}
-
-function KeyPoints({ points, color, radius = 2, fillOpacity = 0.5 }) {
-  return (
-    <g>
-      {points.map(([ x, y ], index) => (
-        <circle
-            key={index}
-            cx={x}
-            cy={y}
-            r={radius}
-            stroke="none"
-            fill={color}
-            fillOpacity={fillOpacity} />
-      ))}
-    </g>
-  );
-}
-
-function Attributes({ sex, age, x1, y1, x2, y2, color, fontSize = 16, opacity = 0.9, strokeWidth = 3 }) {
-  const x = x1;
-  const y = y1 - 7;
-  const label = `Sex: ${sex} / Age: ${age}`;
-  return (
-    <g>
-      <text
-          x={x}
-          y={y}
-          fontSize={fontSize}
-          stroke="white"
-          strokeWidth={strokeWidth}
-          strokeOpacity={opacity}
-          fill="none">
-        {label}
-      </text>
-      <text
-          x={x}
-          y={y}
-          fontSize={fontSize}
-          stroke="none"
-          fill={color}
-          fillOpacity={opacity} >
-        {label}
-      </text>
-    </g>
-  );
-}
-
-function Face({ face, shows }) {
-  const [
-    score,
-    [ x1, y1, x2, y2 ],
-    keyPoints,
-    [ sex, age ],
-  ] = face;
-  const color = {M: "#6666CC", F: "#CC6666"}[sex];
-  return (
-    <g>
-      {!shows.boundingBox ? null :
-        <BoundingBox
-            x1={x1}
-            y1={y1}
-            x2={x2}
-            y2={y2}
-            color={color} />
-      }
-      {!shows.score ? null :
-        <ScoreBar
-            score={score}
-            x1={x1}
-            y1={y1}
-            x2={x2}
-            y2={y2}
-            color={color} />
-      }
-      {!shows.keyPoints ? null :
-        <KeyPoints
-            points={keyPoints}
-            color={"#009900"} />
-      }
-      {!shows.attributes ? null :
-        <Attributes
-            sex={sex}
-            age={age}
-            x1={x1}
-            y1={y1}
-            x2={x2}
-            y2={y2}
-            color={color} />
-      }
-    </g>
-  );
-}
-
-function Faces({ faces, shows }) {
-  if ( faces == null ) return null;
-  return (
-    <g>
-      {faces.map((face, index) => (
-        <Face
-            key={index}
-            face={face}
-            shows={shows} />
-      ))}
-    </g>
-  );
-}
-
-function CheckBox({ id, checked, setChecked, children }) {
-  return (
-    <label>
-      <input
-          type="checkbox"
-          checked={checked[id]}
-          onChange={(e) => {
-            const copied = {...checked};
-            copied[id] = e.target.checked;
-            setChecked(copied);
-          }} />
-      <span style={{cursor: "pointer"}}>
-        {children}
-      </span>
-    </label>
-  );
-}
-
-function Video({ style, src, onTimeChanged }) {
-  const refVideo = useRef(null);
-  const refTimerId = useRef(null);
-  const [paused, setPaused] = useState(true);
-
-  const onChanged = useCallback((event) => {
-    const videoElement = event.target;
-    setPaused(videoElement.paused);
-    onTimeChanged(videoElement.currentTime);
-  }, []);
-
-  const startTimer = useCallback(() => {
-    refTimerId.current = setInterval(() => {
-      if ( refVideo.current != null ) {
-        onTimeChanged(refVideo.current.currentTime);
-      }
-    }, 33);
-  }, [refTimerId, refVideo]);
-
-  const stopTimer = useCallback(() => {
-    if ( refTimerId.current != null ) {
-      clearInterval(refTimerId.current);
-      refTimerId.current = null;
-    }
-  }, [refTimerId]);
-
-  useEffect((x) => {
-    if ( paused ) {
-      stopTimer();
-    } else {
-      startTimer();
-    }
-  }, [paused]);
-
-  return (
-    <video
-        style={style}
-        ref={refVideo}
-        src={src}
-        onLoadStart ={(e) => onChanged(e)}
-        onCanPlay   ={(e) => onChanged(e)}
-        onPause     ={(e) => onChanged(e)}
-        onPlay      ={(e) => onChanged(e)}
-        onSeeked    ={(e) => onChanged(e)}
-        onTimeUpdate={(e) => onChanged(e)}
-        controls={true} />
-  );
-}
-
-function Overlay({ style, width, height, faces, shows }) {
-  return (
-    <svg
-        xmlns="http://www.w3.org/2000/svg"
-        version="1.1"
-        viewBox={`0 0 ${width} ${height}`}
-        style={style}
-        width={width}
-        height={height}>
-      {faces == null ? null : (
-        <Faces
-            faces={faces}
-            shows={shows}/>
-      )}
-    </svg>
-  );
-}
+import CheckBox from "../components/CheckBox";
+import Video from "../components/Video";
+import Overlay from "../components/Overlay";
 
 export default function Page() {
   const [videoMeta, setVideoMeta] = useState(null);
